requirements.txt
gradio
opencv-python-headless
pytesseract
pdf2image
python-docx
Pillow
numpy